+++ packages/test-utils/src/index.ts
@@ -51,4 +51,5 @@ import * as vec3 from 'gl-matrix/vec3';
 import * as vec2 from 'gl-matrix/vec2';
 export { mat4, mat3, quat, vec4, vec3, vec2 };
 
+export * from './create-basic-primitive.js';
 export * from './create-torus-primitive.js';
